reorganize packages for API layer
