+++ avs-api/src/main/java/org/example/age/avs/api/endpoint/AvsApiModule.java
@@ -1,4 +1,4 @@
-package org.example.age.avs.api;
+package org.example.age.avs.api.endpoint;
 
 import dagger.Binds;
 import dagger.Module;
