+++ avs-api/src/testFixtures/java/org/example/age/avs/service/endpoint/test/StubAvsServiceModule.java
@@ -1,13 +1,14 @@
-package org.example.age.test.avs.service;
+package org.example.age.avs.service.endpoint.test;
 
 import dagger.Binds;
 import dagger.Module;
-import org.example.age.avs.api.AvsApi;
-import org.example.age.avs.api.AvsApiModule;
+import io.undertow.server.HttpHandler;
+import org.example.age.avs.api.endpoint.AvsApi;
+import org.example.age.avs.api.endpoint.AvsApiModule;
 import org.example.age.common.api.extractor.builtin.UserAgentAuthMatchDataExtractorModule;
 import org.example.age.common.api.extractor.test.TestAccountIdExtractorModule;
 
-/** Dagger module that binds dependencies for <code>@Named("api") HttpHandler</code>. */
+/** Dagger module that binds dependencies for <code>@Named("api") {@link HttpHandler}</code>. */
 @Module(
         includes = {
             AvsApiModule.class,
