+++ site-service/src/main/java/org/example/age/site/service/SiteServiceModule.java
@@ -7,8 +7,8 @@ import org.example.age.common.api.extractor.AuthMatchDataExtractor;
 import org.example.age.common.service.data.internal.DataMapperModule;
 import org.example.age.common.service.store.PendingStoreFactory;
 import org.example.age.infra.service.client.RequestDispatcherModule;
-import org.example.age.site.api.SiteApi;
-import org.example.age.site.api.SiteApiModule;
+import org.example.age.site.api.endpoint.SiteApi;
+import org.example.age.site.api.endpoint.SiteApiModule;
 import org.example.age.site.service.config.SiteConfig;
 import org.example.age.site.service.config.internal.SiteConfigurerModule;
 import org.example.age.site.service.store.VerificationStore;
