+++ site-service/src/testFixtures/java/org/example/age/test/avs/service/FakeAvsService.java
@@ -7,7 +7,7 @@ import org.example.age.api.Dispatcher;
 import org.example.age.api.JsonSender;
 import org.example.age.api.Sender;
 import org.example.age.api.StatusCodeSender;
-import org.example.age.avs.api.AvsApi;
+import org.example.age.avs.api.endpoint.AvsApi;
 import org.example.age.common.api.data.AuthMatchData;
 import org.example.age.common.service.data.SiteLocation;
 import org.example.age.data.certificate.SignedAgeCertificate;
